+++ ChangeLog.md
@@ -1,6 +1,6 @@
 # Changelog
 ### 07/02/2019 0.18.6
-- Fail the validation when there's schema defined for response body in spec but the example response doesn't provide the body.
+-  Adding support for JSON path & schema path in errors.
 
 ### 06/12/2019 0.18.5